Organized directories. Created the qwixx class and test file
