+++ server/src/models/PlayerClass.ts
