+++ server/src/tests/models/DiceClass.integration.test.ts
