+++ server/src/tests/models/SixSidedDieClass.test.ts
